requirements.txt
requests==2.31.0
pydantic==2.5.3
PyYAML==6.0.1
pytest==7.4.3
py-clob-client==0.19.0
python-dotenv==1.0.0
eth-account>=0.13.0
python-telegram-bot>=20.0
python-dateutil==2.9.0.post0
requests==2.31.0
